Refactor stress experiment tool
